+++ ui/src/views/scenes/Filters.vue
@@ -1,35 +1,10 @@
 <template>
   <div>
-    <div class="field">
-      <label class="label">{{$t("State")}}</label>
-      <div class="control is-expanded">
-        <div class="select is-fullwidth">
-          <select v-model="dlState">
-            <option value="any">{{$t("Any")}}</option>
-            <option value="available">{{$t("Available right now")}}</option>
-            <option value="downloaded">{{$t("Downloaded")}}</option>
-            <option value="missing">{{$t("Not downloaded")}}</option>
-          </select>
-        </div>
-      </div>
-    </div>
+    <SavedSearch/>
 
-    <div class="field">
-      <label class="label">List</label>
-      <b-field>
-        <b-checkbox-button v-model="lists" native-value="watchlist" type="is-primary">
-          <b-icon pack="mdi" icon="calendar-check" size="is-small"/>
-          <span>{{$t("Watchlist")}}</span>
-        </b-checkbox-button>
-        <b-checkbox-button v-model="lists" native-value="favourite" type="is-danger">
-          <b-icon pack="mdi" icon="heart" size="is-small"/>
-          <span>{{$t("Favourite")}}</span>
-        </b-checkbox-button>
-      </b-field>
-    </div>
+    <hr/>
 
-    <label class="label">{{$t("Sort by")}}</label>
-    <div class="field has-addons">
+    <b-field :label="$t('Sort by')" label-position="on-border" :addons="true" class="field-extra">
       <div class="control is-expanded">
         <div class="select is-fullwidth">
           <select v-model="sort">
@@ -46,10 +21,20 @@
           </select>
         </div>
       </div>
-    </div>
+    </b-field>
 
-    <label class="label">Watch status</label>
-    <div class="field has-addons">
+    <b-field :grouped="true" class="field-extra">
+      <b-checkbox-button v-model="lists" native-value="watchlist" type="is-primary">
+        <b-icon pack="mdi" icon="calendar-check" size="is-small"/>
+        <span>{{$t("Watchlist")}}</span>
+      </b-checkbox-button>
+      <b-checkbox-button v-model="lists" native-value="favourite" type="is-danger">
+        <b-icon pack="mdi" icon="heart" size="is-small"/>
+        <span>{{$t("Favourite")}}</span>
+      </b-checkbox-button>
+    </b-field>
+
+    <b-field label="Watched" label-position="on-border" :addons="true" class="field-extra">
       <div class="control is-expanded">
         <div class="select is-fullwidth">
           <select v-model="isWatched">
@@ -59,66 +44,65 @@
           </select>
         </div>
       </div>
-    </div>
+    </b-field>
 
-    <div v-if="Object.keys(filters).length !== 0">
-      <label class="label">Release date</label>
-      <div class="field has-addons">
-        <div class="control is-expanded">
-          <div class="select is-fullwidth">
-            <select v-model="releaseMonth">
-              <option></option>
-              <option v-for="t in filters.release_month" :key="t">{{t}}</option>
-            </select>
-          </div>
-        </div>
-        <div class="control">
-          <button type="submit" class="button is-light" @click="clearReleaseMonth">
-            <b-icon pack="fas" icon="times" size="is-small"></b-icon>
-          </button>
+    <b-field label="Release month" label-position="on-border" :addons="true" class="field-extra">
+      <div class="control is-expanded">
+        <div class="select is-fullwidth">
+          <select v-model="releaseMonth">
+            <option></option>
+            <option v-for="t in filters.release_month" :key="t">{{t}}</option>
+          </select>
         </div>
       </div>
+      <div class="control">
+        <button type="submit" class="button is-light" @click="clearReleaseMonth">
+          <b-icon pack="fas" icon="times" size="is-small"></b-icon>
+        </button>
+      </div>
+    </b-field>
 
+    <hr/>
 
-      <label class="label">Cast</label>
-      <div class="field">
+    <div v-if="Object.keys(filters).length !== 0">
+      <b-field label="Cast" label-position="on-border" class="field-extra">
         <b-taginput v-model="cast" autocomplete :data="filteredCast" @typing="getFilteredCast">
           <template slot-scope="props">{{props.option}}</template>
           <template slot="empty">No matching cast</template>
         </b-taginput>
-      </div>
+      </b-field>
 
-      <label class="label">Site</label>
-      <div class="field">
+      <b-field label="Site" label-position="on-border" class="field-extra">
         <b-taginput v-model="sites" autocomplete :data="filteredSites" @typing="getFilteredSites">
           <template slot-scope="props">{{props.option}}</template>
           <template slot="empty">No matching sites</template>
         </b-taginput>
-      </div>
+      </b-field>
 
-      <label class="label">Tags</label>
-      <div class="field">
+      <b-field label="Tags" label-position="on-border" class="field-extra">
         <b-taginput v-model="tags" autocomplete :data="filteredTags" @typing="getFilteredTags">
           <template slot-scope="props">{{props.option}}</template>
           <template slot="empty">No matching tags</template>
         </b-taginput>
-      </div>
+      </b-field>
 
-      <label class="label">Cuepoint</label>
-      <div class="field">
+      <b-field label="Cuepoint" label-position="on-border" class="field-extra">
         <b-taginput v-model="cuepoint" allow-new>
           <template slot-scope="props">{{props.option}}</template>
           <template slot="empty">No matching cuepoints</template>
         </b-taginput>
-      </div>
+      </b-field>
 
     </div>
   </div>
 </template>
 
 <script>
+  import SavedSearch from "./SavedSearch";
+
   export default {
     name: "Filters",
+    components: {SavedSearch},
     mounted() {
       this.$store.dispatch("sceneList/filters");
     },
@@ -130,7 +114,7 @@
       }
     },
     methods: {
-      reload() {
+      reloadList() {
         this.$router.push({
           name: 'scenes',
           query: {
@@ -155,7 +139,7 @@
       },
       clearReleaseMonth() {
         this.$store.state.sceneList.filters.releaseMonth = "";
-        this.reload();
+        this.reloadList();
       },
     },
     computed: {
@@ -168,36 +152,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.lists = value;
-          this.reload();
-        }
-      },
-      dlState: {
-        get() {
-          return this.$store.state.sceneList.filters.dlState;
-        },
-        set(value) {
-          this.$store.state.sceneList.filters.dlState = value;
-
-          switch (this.$store.state.sceneList.filters.dlState) {
-            case "any":
-              this.$store.state.sceneList.filters.isAvailable = null;
-              this.$store.state.sceneList.filters.isAccessible = null;
-              break;
-            case "available":
-              this.$store.state.sceneList.filters.isAvailable = true;
-              this.$store.state.sceneList.filters.isAccessible = true;
-              break;
-            case "downloaded":
-              this.$store.state.sceneList.filters.isAvailable = true;
-              this.$store.state.sceneList.filters.isAccessible = null;
-              break;
-            case "missing":
-              this.$store.state.sceneList.filters.isAvailable = false;
-              this.$store.state.sceneList.filters.isAccessible = null;
-              break;
-          }
-
-          this.reload();
+          this.reloadList();
         }
       },
       releaseMonth: {
@@ -206,7 +161,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.releaseMonth = value;
-          this.reload();
+          this.reloadList();
         }
       },
       cast: {
@@ -215,7 +170,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.cast = value;
-          this.reload();
+          this.reloadList();
         }
       },
       sites: {
@@ -224,7 +179,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.sites = value;
-          this.reload();
+          this.reloadList();
         }
       },
       tags: {
@@ -233,7 +188,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.tags = value;
-          this.reload();
+          this.reloadList();
         }
       },
       cuepoint: {
@@ -242,7 +197,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.cuepoint = value;
-          this.reload();
+          this.reloadList();
         }
       },
       sort: {
@@ -251,7 +206,7 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.sort = value;
-          this.reload();
+          this.reloadList();
         }
       },
       isWatched: {
@@ -260,9 +215,15 @@
         },
         set(value) {
           this.$store.state.sceneList.filters.isWatched = value;
-          this.reload();
+          this.reloadList();
         }
       },
     }
   }
 </script>
+
+<style scoped>
+  .field-extra {
+    margin-bottom: 1.1em !important;
+  }
+</style>
